* feature/PACS-21_Dynamic_retrieve_with_spreadsheet: (43 commits)
  Cleaned up some warnings
  Working Query test
  Remove a bit of logging
  Synchronize on the poolContainer when processing files
  Anonymizer script now valid XML
  Launch the PACS from Eclipse
  Minor cleanup
  Java side updates
  New name and ID
  Updated CTP script
  Updated test spreadsheet
  General cleanup, adding font-awesome
  Building font-awesome
  Support for fetching, etc.
  Fetch support for the query
  Local font-awesome
  Boolean doFetch
  Query support
  Query controller
  Minor cleanup
  ...

PACS-21 #resolve

+++ src/edu/mayo/qia/pacs/rest/DeviceEndpoint.java
@@ -7,6 +7,9 @@ import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
 import javax.ws.rs.POST;
+import javax.ws.rs.PUT;
+import javax.ws.rs.Path;
+import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
@@ -90,6 +93,32 @@ public class DeviceEndpoint {
     return Response.ok(device).build();
   }
 
+  /** Modify a pool. */
+  @PUT
+  @Path("/{id: [1-9][0-9]*}")
+  @Consumes(MediaType.APPLICATION_JSON)
+  @Produces(MediaType.APPLICATION_JSON)
+  public Response modifyDevice(@PathParam("id") int id, Device update) {
+    // Look up the pool and change it
+    Session session = sessionFactory.openSession();
+    Device device = null;
+    try {
+      session.beginTransaction();
+      device = (Device) session.byId(Device.class).load(id);
+      if (device == null) {
+        return Response.status(Status.NOT_FOUND).entity(new SimpleResponse("message", "Could not load the device")).build();
+      }
+      if (device.pool.poolKey != poolKey) {
+        return Response.status(Status.NOT_FOUND).entity(new SimpleResponse("message", "Could not load the device")).build();
+      }
+      device.update(update);
+      session.getTransaction().commit();
+    } finally {
+      session.close();
+    }
+    return Response.ok(device).build();
+  }
+
   /** Delete a Device. */
   @DELETE
   @Consumes(MediaType.APPLICATION_JSON)
@@ -100,6 +129,9 @@ public class DeviceEndpoint {
     try {
       session.beginTransaction();
       device = (Device) session.byId(Device.class).getReference(device.deviceKey);
+      if (device.pool.poolKey != poolKey) {
+        return Response.status(Status.NOT_FOUND).entity(new SimpleResponse("message", "Could not load the device")).build();
+      }
       session.delete(device);
       session.getTransaction().commit();
     } finally {
